pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.43.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]
